+++ talalprotocol/optimization.py
@@ -1,18 +1,36 @@
 '''
-This file is for optimized versions of the ciphers
+This file is for new or optimized versions of the ciphers
 
+Binary cipher mods:
+handle punctuation and digit inputs
+fix output
 
+Note: Binary cipher might be faster if we implement a binary search
 '''
 
-
+from talalprotocol import *
 import string
+import re
 english_alphabets = list(string.ascii_lowercase)
+string_punctuation = string.punctuation
 
 
 
 
+# message = '1 a b 2'
+# ciphered_version = print(binary_encryption(message))
 
 
+
+
+# x = caesar_cipher('Hi I am gay',1)
+# y = caesar_decipher ('ij j bn',1)
+#
+# x = vigenere_cipher('My name is talal','math')
+# print(x)
+#
+# y = vigenere_decipher(x, 'math')
+# print(y)